Give column I back to the staff who can actually answer it
Napier started filling in CASE DATA column I, "Under supervison?", by reading
the sentence table off the charges page and marking YES when a probation term
was still running on paper as of the clinic date. Iowa Legal Aid tried it and
asked for it back out.

Their reason is not something more parsing fixes. ICOS does not record an early
discharge, records an extension only sometimes, and never carries parole at all,
because parole is corrections rather than the court. Their words: it "is not
going to be reliable because sometime they don't update if probation was pushed
out or ended early", and they establish this per client against the Department
of Corrections website.

So the page cannot answer the question at any level of care, and a wrong YES is
not a cosmetic error: column I is what puts a client's debt into the expungement
sheet's 910.7 restitution column. Blank reads as n/a there, which is what it has
always read and is the honest answer when the only source is stale.

Out: the supervision sentence types, the note explaining the YES, the term
lookup and the write itself. Staying: the sentence table parse, which is a
faithful reading of the page and costs nothing, and the term arithmetic, which
the payment history uses to measure its twelve month window.

The tests keep every wording that used to produce a YES and assert the column
stays empty for all of them, so putting this back fails eight tests rather than
passing quietly into a hearing. The reasoning is in the module docstring and
beside the code, because the parser now sits one line away from a column it
could fill and the next reader deserves to know that was a decision.

900 tests pass.
